Tasks/Task_7/1_student_management_system/requirements.txt
fastapi
uvicorn
sqlmodel
python-jose
python-multipart
passlib[bcrypt]